requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
appnope==0.1.3
asciichartpy==1.5.25
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
blobfile==2.0.1
boto3==1.26.161
botocore==1.29.161
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==2.1.1
click==7.1.2
click-spinner==0.1.10
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
dataclasses-json==0.5.7
debugpy==1.6.5
decorator==5.1.1
dnspython==2.3.0
docopt==0.6.2
email-validator==1.3.1
entrypoints==0.4
executing==1.2.0
fastapi==0.95.0
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.82.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.22.0
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.11.1
idna==3.4
ipykernel==6.20.1
ipython==8.8.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jupyter_client==7.4.8
jupyter_core==5.1.3
kiwisolver==1.4.4
langchain==0.0.118
llama-index==0.4.35
llvd==3.0.1
lxml==4.9.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numpy==1.24.1
oauthlib==3.2.2
openai==0.26.1
orjson==3.8.8
packaging==23.0
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pipreqs==0.4.11
platformdirs==2.6.2
plotly==5.12.0
prompt-toolkit==3.0.36
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodomex==3.17
pydantic==1.10.6
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2022.7.1
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
s3transfer==0.6.1
scikit-learn==1.2.0
scipy==1.10.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
spyql==0.6.0
SQLAlchemy==1.4.47
stack-data==0.6.2
starlette==0.26.1
tabulate==0.8.10
tenacity==8.1.0
threadpoolctl==3.1.0
tiktoken==0.1.2
tokenizers==0.13.2
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
transformers==4.25.1
typing-inspect==0.8.0
typing_extensions==4.4.0
ujson==5.7.0
uritemplate==4.1.1
urllib3==1.26.13
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.18.1
wcwidth==0.2.5
websockets==10.4
yarg==0.1.9
yarl==1.8.2
